Polish: fix bugs, add animations, clean warnings, add README, implement grep
Bug fixes:
- Window slot leak: closing a window now compacts g_wins[] instead of
  leaving dead entries (max 16 was a hard ceiling). win_close_finish().
- Auto-focus: hiding/closed the focused window now reassigns focus to
  the next topmost visible window instead of leaving g_focus = 0.
- Font corruption: fixed 4 glyphs (g, p, q, y) that had 9 bytes in an
  8-byte array, causing visual glitches and compiler warnings.

Animations:
- Opening: slide-up + fade-in (was instant)
- Closing: fade-out + slide-down before destroying the window
- Minimizing: shrink + fade toward the taskbar (restorable on click)

Terminal:
- Implemented 'grep PATTERN [PATH]' — recursive substring search across
  VFS file contents with path:line output. Added strstr() to lib.h.

Polish:
- Warning-free build: fixed misleading-indentation, unused vars/params,
  unterminated strings, sign-compare, and asm mnemonic suffix warnings
  across app_*.c, net.c, paging.c, process.c, gui.c.
- Added comprehensive README.md (features, shortcuts, build/run,
  architecture, screenshots).

Verified via headless QEMU boot to [kikos] desktop with terminal,
grep, ls, and window interactions — no crashes.

+++ kernel/src/net.c
@@ -45,7 +45,7 @@ void net_init(void) {
 }
 
 void net_rx(uint8_t *packet, int len) {
-    if (len < sizeof(eth_header_t)) return;
+    if ((unsigned int)len < sizeof(eth_header_t)) return;
     eth_header_t *eth = (eth_header_t *)packet;
 
     if (memcmp(eth->dst, my_mac, 6) != 0 &&
